fix(review): keep reconciliation cycle-free
Inject host-side effects through hooks while keeping legacy lifecycle integrations in review-status.

This preserves journal reconciliation events without introducing new dependency cycles.
