+++ src/lib/review-status-reconcile.ts
@@ -1,28 +1,82 @@
-import { join } from 'path';
-import { Effect } from 'effect';
-import { notifyPipelineSync } from './pipeline-notifier.js';
-import { needsReviewDispatch } from './review-dispatch-decision.js';
 import { normalizeReviewStatusSync } from './review-status-normalize.js';
 import { upsertReviewStatusSync as dbUpsert } from './overdeck/review-status-sync.js';
 import type { readJournalStatusSync } from './overdeck/review-status-record-sync.js';
+import type { InspectionStatusFields } from './inspection-status.js';
+import type { StrikeLandingStatus } from './strike-landing.js';
+import type { ScopeDriftRecord } from './xbrief/continue-state.js';
 
-type ReviewStatus = Parameters<typeof normalizeReviewStatusSync>[0];
+export interface StatusHistoryEntry {
+  type: 'review' | 'test' | 'merge' | 'inspect' | 'uat' | 'release';
+  status: string;
+  timestamp: string;
+  notes?: string;
+}
+
+export interface BlockerReason {
+  type: 'failing_checks' | 'merge_conflict' | 'unresolved_conversations' | 'changes_requested' | 'draft_pr' | 'not_mergeable';
+  summary: string;
+  details?: string;
+  detectedAt: string;
+}
+
+export interface ReviewStatus extends StrikeLandingStatus, InspectionStatusFields {
+  issueId: string;
+  reviewStatus: 'pending' | 'reviewing' | 'passed' | 'failed' | 'blocked' | 'skipped';
+  testStatus: 'pending' | 'testing' | 'passed' | 'failed' | 'skipped' | 'dispatch_failed';
+  mergeStatus?: 'pending' | 'queued' | 'merging' | 'verifying' | 'merged' | 'failed';
+  releaseStatus?: 'pending' | 'releasing' | 'passed' | 'failed' | 'partial' | 'rolled_back' | 'skipped';
+  uatStatus?: 'pending' | 'testing' | 'passed' | 'failed';
+  uatNotes?: string;
+  verificationStatus?: 'pending' | 'running' | 'passed' | 'failed' | 'skipped';
+  verificationNotes?: string;
+  verificationCycleCount?: number;
+  verificationMaxCycles?: number;
+  reviewNotes?: string;
+  testNotes?: string;
+  mergeNotes?: string;
+  releaseNotes?: string;
+  updatedAt: string;
+  readyForMerge: boolean;
+  autoMerge?: boolean;
+  autoRequeueCount?: number;
+  mergeRetryCount?: number;
+  prUrl?: string;
+  prHeadSha?: string;
+  prNumber?: number;
+  history?: StatusHistoryEntry[];
+  blockerReasons?: BlockerReason[];
+  reviewedAtCommit?: string;
+  lastVerifiedCommit?: string;
+  mergeStep?: string;
+  stuck?: boolean;
+  stuckReason?: string;
+  stuckAt?: string;
+  stuckDetails?: string;
+  reviewSpawnedAt?: string | number;
+  reviewRequestedAt?: string;
+  conflictResolutionDispatchedAt?: string;
+  testRetryCount?: number;
+  reviewRetryCount?: number;
+  recoveryStartedAt?: string;
+  deaconIgnored?: boolean;
+  deaconIgnoredAt?: string;
+  deaconIgnoredReason?: string;
+  scopeDrift?: ScopeDriftRecord;
+  reviewerVerdicts?: Partial<Record<string, {
+    status: 'passed' | 'blocked';
+    atCommit?: string;
+    findingsPath?: string;
+  }>>;
+}
+
 type ReviewStatusJournal = NonNullable<ReturnType<typeof readJournalStatusSync>>;
-export type SetReviewStatusSync = (
-  issueId: string,
-  update: Partial<ReviewStatus>,
-  existing?: ReviewStatus,
-) => ReviewStatus;
 
-export function emitReactiveLifecycleEvent(
-  type: 'review.approved' | 'test.passed',
-  issueId: string,
-): void {
-  try {
-    notifyPipelineSync({ type, issueId });
-  } catch (error) {
-    console.warn(`[review-status] Failed to emit ${type} for ${issueId}:`, error);
-  }
+export interface ReviewStatusReconcileHooks {
+  notifyStatusChanged(issueId: string, status: ReviewStatus): void;
+  deliverReviewVerdictFeedbackHostSide(issueId: string, status: ReviewStatus): Promise<void>;
+  emitReactiveLifecycleEvent(type: 'review.approved' | 'test.passed', issueId: string): void;
+  maybeAutoDispatchReviewHostSide(issueId: string, status: ReviewStatus): void;
+  maybeRecoverTestVerdictHostSide(issueId: string, status: ReviewStatus): void;
 }
 
 export function verificationSatisfied(
@@ -31,9 +85,7 @@ export function verificationSatisfied(
   return status.verificationStatus !== 'failed';
 }
 
-/**
- * PAN-1988: the merge-gate predicate shared by direct writes and journal reconciliation.
- */
+/** PAN-1988: merge-gate predicate shared by direct writes and journal reconciliation. */
 export function reviewGatesPassedSync(
   status: Pick<
     ReviewStatus,
@@ -52,129 +104,12 @@ export function reviewGatesPassedSync(
   );
 }
 
-/**
- * PAN-1988 — deliver review feedback to the work agent from the host when a blocked/failed
- * verdict is reconciled from the journal. Fully best-effort and never throws into the read path.
- */
-export async function deliverReviewVerdictFeedbackHostSide(
-  issueId: string,
-  status: ReviewStatus,
-): Promise<void> {
-  try {
-    const { deliverReviewVerdictFeedback } = await import('./cloister/review-verdict-feedback.js');
-    const result = await Effect.runPromise(deliverReviewVerdictFeedback({
-      issueId,
-      verdict: status.reviewStatus === 'failed' ? 'failed' : 'blocked',
-      notes: status.reviewNotes,
-      prUrl: status.prUrl,
-    }));
-    if (result.agentMessageSent) {
-      console.log(`[review-status] delivered review feedback to the work agent for ${issueId} (host-side)`);
-    }
-  } catch (err) {
-    console.warn(`[review-status] host-side review feedback delivery for ${issueId} did not complete (non-fatal): ${err instanceof Error ? err.message : String(err)}`);
-  }
-}
-
-const reviewDispatchAttemptAt = new Map<string, number>();
-const REVIEW_AUTO_DISPATCH_THROTTLE_MS = 30_000;
-
-/**
- * PAN-1988 auto-heal — re-dispatch review from durable journal intent when the reactive trigger
- * was lost. A 30-second throttle bounds retries for a genuinely gated issue.
- */
-export function maybeAutoDispatchReviewHostSide(issueId: string, status: ReviewStatus): void {
-  if (!needsReviewDispatch({
-    reviewRequestedAt: status.reviewRequestedAt,
-    reviewSpawnedAt: status.reviewSpawnedAt,
-    reviewStatus: status.reviewStatus,
-    mergeStatus: status.mergeStatus,
-  })) return;
-  const last = reviewDispatchAttemptAt.get(issueId) ?? 0;
-  if (Date.now() - last < REVIEW_AUTO_DISPATCH_THROTTLE_MS) return;
-  reviewDispatchAttemptAt.set(issueId, Date.now());
-  void dispatchReviewHostSide(issueId, status.prUrl);
-}
-
-async function dispatchReviewHostSide(issueId: string, prUrl?: string): Promise<void> {
-  try {
-    const { resolveProjectFromIssueSync } = await import('./projects.js');
-    const resolved = resolveProjectFromIssueSync(issueId);
-    if (!resolved) return;
-    const { existsSync } = await import('fs');
-    const workspace = join(resolved.projectPath, 'workspaces', `feature-${issueId.toLowerCase()}`);
-    if (!existsSync(workspace)) return;
-    let branch = `feature/${issueId.toLowerCase()}`;
-    try {
-      const { promisify } = await import('util');
-      const { exec } = await import('child_process');
-      const execAsync = promisify(exec);
-      const { stdout } = await execAsync('git branch --show-current', { cwd: workspace, encoding: 'utf-8' });
-      branch = stdout.trim() || branch;
-    } catch { /* non-fatal — fall back to the conventional branch name */ }
-    const { spawnReviewRoleForIssue } = await import('./cloister/review-agent.js');
-    const result = await Effect.runPromise(spawnReviewRoleForIssue({ issueId, workspace, branch, ...(prUrl ? { prUrl } : {}) }));
-    if (result.success) {
-      if (result.message?.startsWith('Review already in progress')) {
-        console.log(`[review-status] review dispatch for ${issueId}: already in progress — no-op (host-side)`);
-      } else {
-        console.log(`[review-status] auto-dispatched review for ${issueId} from durable journal intent (host-side)`);
-      }
-    }
-  } catch (err) {
-    console.warn(`[review-status] host-side review auto-dispatch for ${issueId} did not complete (non-fatal): ${err instanceof Error ? err.message : String(err)}`);
-  }
-}
-
-const testVerdictRecoveryAt = new Map<string, number>();
-const TEST_VERDICT_RECOVERY_THROTTLE_MS = 60_000;
-
-export function maybeRecoverTestVerdictHostSide(
-  issueId: string,
-  status: ReviewStatus,
-  setReviewStatus: SetReviewStatusSync,
-): void {
-  if (status.reviewStatus !== 'passed') return;
-  if (status.mergeStatus === 'merged' || status.readyForMerge) return;
-  if (status.testStatus !== 'testing' && status.testStatus !== 'pending' && status.testStatus !== 'dispatch_failed') return;
-  const last = testVerdictRecoveryAt.get(issueId) ?? 0;
-  if (Date.now() - last < TEST_VERDICT_RECOVERY_THROTTLE_MS) return;
-  testVerdictRecoveryAt.set(issueId, Date.now());
-  void recoverTestVerdictHostSide(issueId, setReviewStatus);
-}
-
-async function recoverTestVerdictHostSide(
-  issueId: string,
-  setReviewStatus: SetReviewStatusSync,
-): Promise<void> {
-  try {
-    const { resolveProjectFromIssueSync } = await import('./projects.js');
-    const resolved = resolveProjectFromIssueSync(issueId);
-    if (!resolved) return;
-    const { existsSync } = await import('fs');
-    const workspace = join(resolved.projectPath, 'workspaces', `feature-${issueId.toLowerCase()}`);
-    if (!existsSync(workspace)) return;
-    const { readTestVerdictArtifact } = await import('./cloister/test-verdict.js');
-    const artifact = readTestVerdictArtifact(workspace);
-    if (!artifact) return;
-    setReviewStatus(issueId, {
-      testStatus: artifact.status,
-      testNotes: artifact.notes ?? `Recovered from .pan/test/result.json (${artifact.status}) — the test agent wrote the verdict but never signaled`,
-    });
-    console.log(`[review-status] recovered unsignaled test verdict for ${issueId}: ${artifact.status} (host-side, from .pan/test/result.json)`);
-  } catch (err) {
-    console.warn(`[review-status] host-side test verdict recovery for ${issueId} did not complete (non-fatal): ${err instanceof Error ? err.message : String(err)}`);
-  }
-}
-
-/**
- * Reconcile a newer canonical journal verdict into the SQLite cache and re-emit the transition.
- */
+/** Reconcile a newer canonical journal verdict into the SQLite cache. */
 export function reconcileJournalIntoCacheSync(
   issueId: string,
   dbStatus: ReviewStatus | null,
   journal: ReviewStatusJournal,
-  setReviewStatus: SetReviewStatusSync,
+  hooks: ReviewStatusReconcileHooks,
 ): ReviewStatus {
   const merged: ReviewStatus = {
     ...(dbStatus ?? {
@@ -200,7 +135,7 @@ export function reconcileJournalIntoCacheSync(
     dbUpsert(reconciled);
     // PAN-2988 — the reconcile changed the effective status; the read model only
     // advances via events, so emit the transition the lost write never produced.
-    notifyPipelineSync({ type: 'status_changed', issueId, status: reconciled });
+    hooks.notifyStatusChanged(issueId, reconciled);
   } catch {
     // Read-only DB (a sandboxed reader) — the host reconciles when it reads. Non-fatal.
   }
@@ -208,23 +143,23 @@ export function reconcileJournalIntoCacheSync(
   const wasBlocked = dbStatus?.reviewStatus === 'blocked' || dbStatus?.reviewStatus === 'failed';
   const nowBlocked = reconciled.reviewStatus === 'blocked' || reconciled.reviewStatus === 'failed';
   if (nowBlocked && !wasBlocked) {
-    void deliverReviewVerdictFeedbackHostSide(issueId, reconciled);
+    void hooks.deliverReviewVerdictFeedbackHostSide(issueId, reconciled);
   }
 
   const wasReviewPassed = dbStatus?.reviewStatus === 'passed';
   const nowReviewPassed = reconciled.reviewStatus === 'passed';
   if (nowReviewPassed && !wasReviewPassed && reconciled.testStatus === 'pending') {
     console.log(`[review-status] reconcile: review.approved for ${issueId} (host-owned handoff — sandboxed agent verdict from journal)`);
-    emitReactiveLifecycleEvent('review.approved', issueId);
+    hooks.emitReactiveLifecycleEvent('review.approved', issueId);
   }
   const wasTestPassed = dbStatus?.testStatus === 'passed';
   const nowTestPassed = reconciled.testStatus === 'passed';
   if (nowTestPassed && !wasTestPassed) {
     console.log(`[review-status] reconcile: test.passed for ${issueId} (host-owned handoff — sandboxed agent verdict from journal)`);
-    emitReactiveLifecycleEvent('test.passed', issueId);
+    hooks.emitReactiveLifecycleEvent('test.passed', issueId);
   }
 
-  maybeAutoDispatchReviewHostSide(issueId, reconciled);
-  maybeRecoverTestVerdictHostSide(issueId, reconciled, setReviewStatus);
+  hooks.maybeAutoDispatchReviewHostSide(issueId, reconciled);
+  hooks.maybeRecoverTestVerdictHostSide(issueId, reconciled);
   return reconciled;
 }
